Create Top K
The top_k parameter limits the model's choices to the k most likely next token. When top_k is set to 1, the model selects the most probable token and in this case its outputs, given the exact same prompt, will always be the same. We refer to this as greedy decoding.

When top_k is set to greater than 1, the model can consider more than one possible next token, not just what it evaluates as the most likely.

So far we have been working with a default value of top_k set to 1.

Here's some toy code to help with your intuition of top_k. Imagine these are the candidate next words generated by the LLM.